requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
distro==1.9.0
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.28
multidict==6.6.4
openai==1.107.3
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pdf2image==1.17.0
pillow==11.3.0
propcache==0.3.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.8.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.25.0
